backend/pyproject.toml
[project]
name = "pack-me-up"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi-sqlalchemy>=0.2.1",
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.1",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


[dependency-groups]
dev = [
    "ruff>=0.11.5",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.27.0",
    "pre-commit>=3.6.0",
]

[tool.isort]
profile = "black"
skip = [
    ".venv",
    "venv",
    "node_modules"
]

[tool.mypy]
python_version = "3.11"
allow_untyped_globals = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F401",
    "F403",
    "F404",
    "F406",
    "F407",
    "F821",
    "F822",
    "F823",
    "F842",
    "D",
    "D404",
    "D401",
    "D403",
    "D417"
]
ignore = [
    "E402",
    "E501",
    "E713",
    "E714",
    "E721",
    "E722",
    "E731",
    "E741",
    # Ignore mandatory docstrings
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    # Ignore whitespace issuses around the docstring
    "D202",
    "D203",
    # Allow starting multi-line docstring on both the 1st or 2nd line
    "D212",
    "D213",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.sources]
fastapi-sqlalchemy = { path = "../../../fastapi-sqlalchemy/dist/FastAPI_SQLAlchemy-0.2.1.post1-py3-none-any.whl" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "unit: mark a test as a unit test",
    "integration: mark a test as an integration test",
    "slow: mark a test as slow",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
